+++ test/spec/resource_collection_agent_test.js
@@ -1,12 +1,12 @@
 "use strict";
 
-describe('ResourceAgent', function() {
+describe('ResourceCollectionAgent', function() {
   describe('constructor', function() {
     it('should take resource name and its client as arguments', function() {
       var form, agent;
 
-      form = { id: 123 };
-      agent = new Cape.ResourceAgent('user', form);
+      form = {};
+      agent = new Cape.ResourceCollectionAgent('user', form);
 
       expect(agent.resourceName).to.equal('user');
       expect(agent.client).to.equal(form);
@@ -15,51 +15,52 @@ describe('ResourceAgent', function() {
     it('should take an object (options) as the third argument', function() {
       var form, options, agent;
 
-      form = { id: 123 };
+      form = {};
       options = { pathPrefix: '/api/' }
-      agent = new Cape.ResourceAgent('user', form, options);
+      agent = new Cape.ResourceCollectionAgent('user', form, options);
 
       expect(agent.options).to.equal(options);
     })
   })
 
-  describe('#memberPath', function() {
+  describe('#collectionPath', function() {
     it('should return standard values', function() {
       var form, agent;
 
-      form = { id: 123 };
-      agent = new Cape.ResourceAgent('user', form);
+      form = {};
+      agent = new Cape.ResourceCollectionAgent('user', form);
 
-      expect(agent.memberPath()).to.equal('/users/123');
+      expect(agent.collectionPath()).to.equal('/users');
     })
 
     it('should add prefix to the paths', function() {
       var form, agent;
 
-      form = { id: 123 };
-      agent = new Cape.ResourceAgent('user', form, { pathPrefix: '/api/' });
+      form = {};
+      agent = new Cape.ResourceCollectionAgent('user', form, { pathPrefix: '/api/' });
 
-      expect(agent.memberPath()).to.equal('/api/users/123');
+      expect(agent.collectionPath()).to.equal('/api/users');
     })
   })
 
   describe('#init', function() {
     it('should go through a fetch api chain', function() {
-      var form, agent, spy1, spy2, spy3;
+      var form, agent, spy1, spy2, spy3, users;
 
-      form = { id: 123 };
-      agent = new Cape.ResourceAgent('user', form);
+      form = {};
+      agent = new Cape.ResourceCollectionAgent('user', form);
 
       spy1 = sinon.spy();
       spy2 = sinon.spy();
       spy3 = sinon.spy();
+      users = [{ id: 1 }]
       sinon.stub(global, 'fetch', function(path, options) {
         return {
           then: function(callback1) {
             callback1.call(this, { json: spy1 })
             return {
               then: function(callback2) {
-                callback2.call(this, { user: {} })
+                callback2.call(this, { users: users })
                 return {
                   catch: function(callback3) {
                     callback3.call(this, new Error(''))
@@ -75,17 +76,18 @@ describe('ResourceAgent', function() {
       expect(spy1.called).to.be.true;
       expect(spy2.called).to.be.true;
       expect(spy3.called).to.be.true;
+      expect(agent.objects).to.equal(users);
 
       global.fetch.restore();
     })
   })
 
-  describe('#update', function() {
+  describe('#create', function() {
     it('should go through a fetch api chain', function() {
       var form, agent, spy1, spy2, spy3;
 
       form = { id: 123, paramsFor: function() { return {} } };
-      agent = new Cape.ResourceAgent('user', form);
+      agent = new Cape.ResourceCollectionAgent('user', form);
 
       spy1 = sinon.spy();
       spy2 = sinon.spy();
@@ -108,44 +110,7 @@ describe('ResourceAgent', function() {
         }
       });
 
-      agent.update(spy2, spy3);
-      expect(spy1.called).to.be.true;
-      expect(spy2.called).to.be.true;
-      expect(spy3.called).to.be.true;
-
-      global.fetch.restore();
-    })
-  })
-
-  describe('#destroy', function() {
-    it('should go through a fetch api chain', function() {
-      var form, agent, spy1, spy2, spy3;
-
-      form = { id: 123 };
-      agent = new Cape.ResourceAgent('user', form);
-
-      spy1 = sinon.spy();
-      spy2 = sinon.spy();
-      spy3 = sinon.spy();
-      sinon.stub(global, 'fetch', function(path, options) {
-        return {
-          then: function(callback1) {
-            callback1.call(this, { json: spy1 })
-            return {
-              then: function(callback2) {
-                callback2.call(this, {})
-                return {
-                  catch: function(callback3) {
-                    callback3.call(this, new Error(''))
-                  }
-                }
-              }
-            }
-          }
-        }
-      });
-
-      agent.destroy(spy2, spy3);
+      agent.create(spy2, spy3);
       expect(spy1.called).to.be.true;
       expect(spy2.called).to.be.true;
       expect(spy3.called).to.be.true;